+++ src/d-responsive-layout/DRow.vue
@@ -1,8 +1,15 @@
 <template>
   <d-box
     class="ui-responsive__row"
-    :class="{ [`align__${alignment}`]: alignment, stretch }"
-    :style="{ '--gutter-x': computedGutter.x, '--gutter-y': computedGutter.y }"
+    :class="{
+      [`align__${alignment}`]: stretch !== undefined ? false : alignment,
+      stretch,
+      noColPadding,
+    }"
+    :style="{
+      '--d-gutter-x': computedGutter.x,
+      '--d-gutter-y': computedGutter.y,
+    }"
   >
     <slot></slot>
   </d-box>
@@ -31,6 +38,9 @@ const props = defineProps({
   stretch: {
     type: Boolean,
   },
+  noColPadding: {
+    type: Boolean,
+  },
 });
 
 const computedGutter = computed(() => {
@@ -49,7 +59,6 @@ const computedGutter = computed(() => {
   display: flex;
   flex-wrap: wrap;
   width: 100%;
-  gap: var(--gutter-x);
   &.stretch {
     align-items: stretch;
   }
@@ -64,7 +73,11 @@ const computedGutter = computed(() => {
     width: 100%;
     max-width: 100%;
     //padding-right: var(--gutter-x);
-    margin-top: var(--gutter-y);
+    margin-top: var(--d-gutter-y);
+  }
+  &:not(.noColPadding) > * {
+    padding-right: calc(var(--d-gutter-x) * 0.5);
+    padding-left: calc(var(--d-gutter-x) * 0.5);
   }
   &.align__top {
     align-items: flex-start;