+++ changelog.md
@@ -1,3 +1,5 @@
+- **2nd August, 2023 - Eric McWinNEr 😴**
+  - Made bugfixes to the responsive layout
 - **1st August, 2023 - Eric McWinNEr 😴**
   - Made minor update to the pin input, tag input and tag dropdown
   - Made the right icon of the select field customizable